add edit task
VSCode gives extremely useful debugging information for Rust (much as it does
for Go) IF the root folder is the folder containing the relevant Cargo.toml
file. Maybe there's a way we can place one in the root folder to cover all days'
projects (as with the go.mod) but for now, we'll take this approach.

To use, type 'task 2017:edit -- 01' (or whatever day you are working on.)