day 13, part 2 -> 3833504
My first attempt was on the right track, but for each step it was
repeating a lot of calculation; the time taken was growing
exponentially! Adding the save_state/restore_state reduced that
overhead. If the answer had been a lower number it would have been
overkill, but ultimately (at 60000) it became apparent that I needed to
improve my approach!