brute force method is not going to work
our initial dance has 10000 steps. It takes roughly 4 seconds
to process 1000 iterations, so 1 billion iterations would take
around 1100 hours to process. Ain't nobody got time for that.

I'm thinking if we analyse the output after 1000 steps, and
generate the simplest possible dance that would generate that
from the input, we'll then have a, say, 100 step dance that,
when run 1 million times, should take just a few seconds to
complete. We can test it on the output after 200 steps, 3000
steps, etc...